feature-sel: split up the moses scorer into two
That one file seemed to contain two totally unrelated scorers,
so I split them into two files.
