+++ opencog/learning/feature-selection/scorers/moses_optim.h
@@ -1,4 +1,4 @@
-/** moses_based_scorer.h --- 
+/** scorers/moses_optim.h --- 
  *
  * Copyright (C) 2011 OpenCog Foundation
  *
@@ -21,8 +21,8 @@
  */
 
 
-#ifndef _OPENCOG_MOSES_BASED_SCORER_H
-#define _OPENCOG_MOSES_BASED_SCORER_H
+#ifndef _OPENCOG_FS_SCORERS_OPTIM_H
+#define _OPENCOG_FS_SCORERS_OPTIM_H
 
 #include <opencog/util/numeric.h>
 
@@ -43,40 +43,6 @@ using namespace combo;
  */
 std::set<arity_t> get_feature_set(const field_set& fields,
                                   const instance& inst);
-
-/**
- * Wrapper to use moses scoring precision (see
- * opencog/learning/moses/moses/scoring.h).
- *
- * That wrapper uses the method best_possible_score() given a certain
- * feature set. And therefore attempts to maximize the best possible
- * score one would get (w.r.t. some fitness function) given the
- * feature set being evaluated.
- */
-template<typename FeatureSet>
-struct pre_scorer : public unary_function<FeatureSet, double> {
-    pre_scorer(const CTable& ctable,
-               float penalty = 1.0f,
-               float min_activation = 0.5f,
-               float max_activation = 1.0f,
-               bool positive = true)
-        : _ctable(ctable), _penalty(penalty),
-          _min_activation(min_activation), _max_activation(max_activation),
-          _positive(positive) {}
-
-    double operator()(const FeatureSet& fs) const {
-        // filter the ctable
-        CTable filtered_ctable = _ctable.filtered(fs);
-        // create the scorer
-        precision_bscore sc(filtered_ctable, _penalty,
-                            _min_activation, _max_activation, _positive);
-        return boost::accumulate(sc.best_possible_bscore(), 0.0);
-    }
-protected:
-    const CTable& _ctable;
-    float _penalty, _min_activation, _max_activation;
-    bool _positive;
-};
 
 /**
  * Wrapper to use a feature set scorer with MOSES's optimization
@@ -114,4 +80,4 @@ struct deme_based_scorer : public iscorer_base
 
 } // ~namespace opencog
 
-#endif // _OPENCOG_MOSES_BASED_SCORER_H
+#endif // _OPENCOG_FS_SCORERS_OPTIM_H
