+++ opencog/learning/feature-selection/scorers/mutual_info.h
@@ -21,8 +21,8 @@
  */
 
 
-#ifndef _OPENCOG_FEATURE_SCORER_H
-#define _OPENCOG_FEATURE_SCORER_H
+#ifndef _OPENCOG_FEATURE_SCORERS_MI_H
+#define _OPENCOG_FEATURE_SCORERS_MI_H
 
 #include <opencog/util/numeric.h>
 #include <opencog/comboreduct/table/table.h>
@@ -122,4 +122,4 @@ struct MICScorerCTable : public std::unary_function<FeatureSet, double>
 
 } // ~namespace opencog
 
-#endif // _OPENCOG_FEATURE_SCORER_H
+#endif // _OPENCOG_FEATURE_SCORERS_MI_H
